[authorize,#11][s]: remove md5, name and type fields from return properties of authorize API as not needed - fixes #11.
Why?

There is no reason you have to supply the file name or type in the upload - it's optional. By having it in return value we require it in the input and get a 500 error if we don't supply e.g. hame (just bitten by this).

Furthermore, user supplied them in the first place so must know them - so returning them serves no purpose.

Going forward we want the input API to only have those fields need to construct the key using STORAGE_PATH_PATTERN